Implement print_table() function
Implement the print_table() function that prints a table with
a ship amount to be placed. The table is printed next to the board
at the placement phase of the game. The location of the table is
controlled by the go_to_point() function that moves a cursor to
the right-hand side of the board.
Implement the clear() function that uses the 'os' module to run
'clear' command.