Change the structure of the ship_stats dictionary
Every ship is represented as a single dictionary containing the
coordinates of its modules, the coordinates of shotted modules,
and its length. All ships are stored in a single dictionary
'ship_stats' which every key corresponds to the ship type. The
key values are lists containing dictionaries of every ship of its
kind.