Implement new version of print_board() function
The new version of print_board() function allows specifying the
coordinates of the upper-left pixel of the board. The function
also receives a ‘player’ dictionary that is used to print the
player’s name above the board.
Implement print_boards() function. The function uses the modified
version of the print_board() function.