Add new feature - print messages
Modify the function shot() so that it returns a letter
corresponding to the shot result.

Implement the shot_msg() function that returns a message that
corresponds to the shot result.
Implement the coordinates_index_to_string() function that returns
a descriptive name of coordinates, e.g. A2.

Modify the battleship_game() function by including the above
functions.