Modify the main game loop graphical interface
Add some constants to functions: print_boards() and
battleship_game(). The constants are used to correctly align both
boards and tables with ship stats.

Implement the update_ship() function that returns updated ‘ships’
dictionary. The updates dictionary is used as an argument for the
print_table() function implemented in the battleship_game()
function.