Modify the UI
Add constants that store different types of marks and their colors.

Implement the function board_row_color_parser() that returns a list
of marks with color description.

Correct the part of the function ai_place_ship() that resets the
references of dictionaries: ‘board’ and ‘ship_stats’. Previously,
the function was overwriting the references.