Correct the function get_player_target_for_shot()
Solve the bug that was crashing the game when a player input wrong
coordinates at the shotting phase.

Improve displaying messages, so no additional lines are added to
the terminal screen.

Implement the function clear_line() that removes all characters
from the cursor position to the end of the line.